Move SectionTemplateTest to Tag service
